RES-1152 Complete library and tests on it.

+++ Mds.Libraries.CursiveAmount/Interfaces/ICurrencyFormatter.cs
@@ -2,7 +2,7 @@
 {
     public interface ICurrencyFormatter
     {
-        string Build(int amount);
+        string Build(long amount);
         string Build(double amount);
     }
 }
